src/web_api/requirements.txt
psycopg2-binary
types-psycopg2
pydantic
pydantic[email]
python-dotenv
redis